Began splitting up the codebase
