+++ package.json
@@ -4,8 +4,8 @@
   "description": "A library that implements Nix language semantics in JS.",
   "main": "src/lib.js",
   "scripts": {
-    "build": "tsc -p . --outDir dist",
-    "build-watch": "tsc -w -p . --outDir dist",
+    "build": "tsup ./src/lib.ts --format esm --sourcemap",
+    "build-watch": "yarn build --watch",
     "fmt": "prettier --write .",
     "fmt-check": "prettier --check .",
     "test": "jest --color",
@@ -26,6 +26,7 @@
     "jest": "29.7.0",
     "prettier": "3.2.5",
     "ts-jest": "29.1.2",
+    "tsup": "^8.0.2",
     "typescript": "5.3.3"
   },
   "files": [
